requirements.txt
bs4
pandas
attrs
requests
lxml
nflgame-redux
pyarrow
tqdm